Changed the return value for IntegerCVParam and LongCVParam to be the correct types

+++ src/main/java/com/alanmrace/jimzmlparser/mzML/IntegerCVParam.java
@@ -35,7 +35,7 @@ public class IntegerCVParam extends CVParam {
 		this.units = cvParam.units;
 	}
 	
-	public double getValue() {
+	public int getValue() {
 		return value;
 	}
